Add linkdep.sh and refactor test.sh to be compatible with it

+++ bin/test.sh
@@ -1,8 +1,5 @@
 #!/bin/bash
 
-# Stop testing later steps if an earlier step fails.
-set -e
-
 # Save the arguments for use within the functions below.
 ARGS=("$@")
 
@@ -16,26 +13,48 @@ initialize() {
 }
 
 link_vendor_in_cwd() {
-    if [ -d vendor ]; then
-        # In case someone has the old glide vendor directory lying around.
-        rm -rf vendor
-    else
-        # Ensure there is no vendor file or symbolic link here.
-        rm -f vendor
-    fi
+    REPO="$1"
+
+    unlink_vendor_in_cwd "$REPO"
 
     # Allow go test to find all the dependencies in the commands vendor directory.
+    echo linking vendor directory in "$REPO"
     ln -s "$COMMANDS_DIR"/vendor vendor
+
+    # Avoid circular references when testing the given repo.
+    DIR="$COMMANDS_DIR"/vendor/github.com/oneiro-ndev/"$REPO"
+    if [ -e "$DIR" ]; then
+        echo moving away "$REPO" subdirectory in vendor directory
+        rm -rf "$DIR-backup"
+        mv "$DIR" "$DIR-backup"
+    fi
 }
 
 unlink_vendor_in_cwd() {
-    # Clean up symbolic link to commands vendor directory
-    rm -f vendor
+    REPO="$1"
+
+    DIR="$COMMANDS_DIR"/vendor/github.com/oneiro-ndev/"$REPO"
+    if [ -e "$DIR-backup" ]; then
+        echo moving back "$REPO" subdirectory in vendor directory
+        rm -rf "$DIR"
+        mv "$DIR-backup" "$DIR"
+    fi
+
+    if [ -e vendor ]; then
+        echo removing vendor from "$REPO"
+        if [ -d vendor ]; then
+            # In case someone has the old glide vendor directory lying around.
+            rm -rf vendor
+        else
+            # Ensure there is no vendor file or symbolic link here.
+            rm -f vendor
+        fi
+    fi
 }
 
 test_chaos() {
     cd "$CHAOS_DIR"
-    link_vendor_in_cwd
+    link_vendor_in_cwd chaos
 
     chaosintegration=0
     for arg in "${ARGS[@]}"; do
@@ -45,16 +64,18 @@ test_chaos() {
         fi
     done
     if [ "$chaosintegration" != 1 ]; then
+        echo
         echo testing chaos
         go test ./...
+        echo
     fi
 
-    unlink_vendor_in_cwd
+    unlink_vendor_in_cwd chaos
 }
 
 test_ndau() {
     cd "$NDAU_DIR"
-    link_vendor_in_cwd
+    link_vendor_in_cwd ndau
 
     ndauintegration=0
     for arg in "${ARGS[@]}"; do
@@ -64,8 +85,10 @@ test_ndau() {
         fi
     done
     if [ "$ndauintegration" != 1 ]; then
+        echo
         echo testing ndau
         go test ./...
+        echo
     else
         # Integration tests require that the node group is running.
         "$CMDBIN_DIR"/run.sh
@@ -73,16 +96,18 @@ test_ndau() {
         # Sleep one more second so that tendermint has a chance to become ready.
         sleep 1
 
+        echo
         echo testing ndau integration
         NDAU_RPC=http://localhost:$TM_NDAU_RPC_PORT
         CHAOS_RPC=http://localhost:$TM_CHAOS_RPC_PORT
         go test ./pkg/ndauapi/routes/... -integration -ndaurpc="$NDAU_RPC" -chaosrpc="$CHAOS_RPC"
+        echo
 
         # We forced-ran for integration tests, so we might as well kill automatically too.
         "$CMDBIN_DIR"/kill.sh
     fi
 
-    unlink_vendor_in_cwd
+    unlink_vendor_in_cwd ndau
 }
 
 test_all() {
